+++ src/main/java/de/ummahuesla/doiwanttolivehere/collectors/NearbyDoctorsCollector.java
@@ -7,8 +7,8 @@ import de.ummahuesla.doiwanttolivehere.model.Score;
 public class NearbyDoctorsCollector extends Collector {
 
 	@Override
-	public Score getScore(long lat, long lon) {
-		return Score.create("Ärtzte in der Nähe", ThreadLocalRandom.current().nextInt(scoreMin, scoreMax+ 1));
+	public Score getScore(Double lat, Double lon) {
+		return Score.create("Ärtzte in der Nähe", ThreadLocalRandom.current().nextDouble(scoreMin, scoreMax+ 1));
 	}
 
 }
