+++ src/main/java/de/ummahuesla/doiwanttolivehere/collectors/NearbySupermarketsCollector.java
@@ -7,8 +7,8 @@ import de.ummahuesla.doiwanttolivehere.model.Score;
 public class NearbySupermarketsCollector extends Collector {
 
 	@Override
-	public Score getScore(long lat, long lon) {
-		return Score.create("Supermärkte in der Nähe", ThreadLocalRandom.current().nextInt(scoreMin, scoreMax+ 1));
+	public Score getScore(Double lat, Double lon) {
+		return Score.create("Supermärkte in der Nähe", ThreadLocalRandom.current().nextDouble(scoreMin, scoreMax+ 1));
 	}
 
 }
