+++ src/main/java/de/ummahuesla/doiwanttolivehere/collectors/SunHoursPerYearCollector.java
@@ -7,8 +7,8 @@ import de.ummahuesla.doiwanttolivehere.model.Score;
 public class SunHoursPerYearCollector extends Collector {
 
 	@Override
-	public Score getScore(long lat, long lon) {
-		return Score.create("Sonnenstunden pro Jahr", ThreadLocalRandom.current().nextInt(scoreMin, scoreMax + 1));
+	public Score getScore(Double lat, Double lon) {
+		return Score.create("Sonnenstunden pro Jahr", ThreadLocalRandom.current().nextDouble(scoreMin, scoreMax + 1));
 	}
 
 }
